refactor(plugin): split importer.ts into focused modules
Pure moves, no logic changes: collection-specs, variable-sync, fonts,
text-styles, effect-styles, variable-values, and color modules.
importer.ts stays as the orchestrator; resolveCompositeValue moves to
resolver.ts and toCssColor joins parseColorValue in color.ts.
